Implement END_FINALLY for Py3
In python 3, END_FINALLY is responsible for a little more work.  If a context manager should suppress an error, WITH_CLEANUP doesn't pop the exception off the stack.  Instead, it pushes a new 'why' code, WHY_SILENCED, onto the stack, and leaves the popping to END_FINALLY.

Context managers always end with an END_FINALLY in py3, even though they don't start with a SETUP_FINALLY.  Context managers in py2 don't end with END_FINALLY, so this code won't be executed with context managers in py2.