Remove "dis" module from pyvm2.py completely...
It was preventing cross version interpreters from working.

For example byte 3.7 and 2.6 have a `LOAD_ATTR` operation,
however the opcode number is different between the two, even
though they do semantically the same thing. (Thanks, Python!)

So using 3.7's dis to figure out how to interpret the operand of
LOAD_ATTR will fail because based on the opocde number it won't
see this opcode number as being an argument that works with names.