Start to make sense of VMRuntime errors...
these are RuntimeErrors in the program and they should be separated
errors and bugs in the VM, or in the setup of the VM.

However we have this monkey on our shoulder: test/vmtest.py

It has some great features, and some not-so-great features.
For now we, have a flag in VirtualMachine just to coddle to
test/vmtest.py.

Todo: fix up Python 2.x RuntimeError tracebacks and limit
3.x RuntimeEror tracebacks to eliminate the VM portion.