Fill out interpreter inspect more...
This is most needed in cross-version interpretation, but it is also
needed within a single Python version because interpreter types are not
the same thing as the built-in types.

More work is needed here, but this makes some progress.

Note: a lot of stdlib tests have become broken. This is probably good
because it reflects that we are honestly interpreting more and not
running interpreter code as builtin Python functions (e.g. eval and
exec used to work this way.)